+++ requirements.txt
@@ -1,0 +1,10 @@
+numpy
+scipy
+torch       
+paho-mqtt
+pyttsx3
+SpeechRecognition
+openai
+tk 
+pyaudio
+# use pip3 install --user -r requirements.txt
